Otimização Api
