+++ src/controllers/geografia.php
@@ -1,0 +1,56 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: jeanfreitas
+ * Date: 24/02/2018
+ * Time: 08:08
+ */
+
+use Symfony\Component\HttpFoundation\Request;
+
+$app->get('api/paises', function (Request $request) use ($app) {
+    /** @var \Doctrine\DBAL\Connection $db */
+    $db = $app['db'];
+    $query = "SELECT * FROM paises";
+    $id = (int)$request->get('id');
+    $params = [];
+    if ($id > 0) {
+        $query .= " WHERE id = ?";
+        array_push($params, $id);
+    }
+    $result = $db->fetchAll($query, $params);
+    return $app->json($result);
+});
+
+$app->get('api/estados', function (Request $request) use ($app) {
+    /** @var \Doctrine\DBAL\Connection $db */
+    $db = $app['db'];
+    $query = "SELECT * FROM estados";
+    $id = (int)$request->get('id');
+    $params = [];
+    if ($id > 0) {
+        $query .= " WHERE id = ?";
+        array_push($params, $id);
+    }
+    $result = $db->fetchAll($query, $params);
+    return $app->json($result);
+});
+
+$app->get('api/cidades', function (Request $request) use ($app) {
+    /** @var \Doctrine\DBAL\Connection $db */
+    $db = $app['db'];
+    $query = "SELECT * FROM cidades";
+    $id = (int)$request->get('id');
+    $id_estado = (int)$request->get('uf');
+    $params = [];
+    if ($id > 0) {
+        $query .= " WHERE id = ?";
+        array_push($params, $id);
+    }
+    if ($id_estado > 0) {
+        $query .= " WHERE id_estado = ?";
+        array_push($params, $id_estado);
+    }
+    $result = $db->fetchAll($query, $params);
+    return $app->json($result);
+});
